refactored structure for solution
