+++ src/App.js
@@ -1,5 +1,5 @@
 import React from "react";
-import NavBar from "./NavBar";
+import NavBar from "./components/NavBar";
 import { Outlet } from "react-router-dom"
 
 function App() {
