+++ src/data.js
@@ -50,7 +50,7 @@ const directors = [
   },
   {
     name: "Edward Zwick",
-    movies: ["Jack Reacher: Never Go Back", "Blood Diamon", "The Siege"],
+    movies: ["Jack Reacher: Never Go Back", "Blood Diamond", "The Siege"],
   },
 ];
