+++ src/pages/ErrorPage.js
@@ -1,4 +1,4 @@
-import React from 'react'
+
 
 function ErrorPage() {
   return (
