+++ src/pages/Home.js
@@ -1,5 +1,4 @@
 import React from "react";
-import NavBar from "./NavBar";
 
 function Home() {
   return <div>
